+++ common/build.go
@@ -353,7 +353,10 @@ func (b *Build) Run(globalConfig *Config, trace JobTrace) (err error) {
 	var executor Executor
 
 	b.logger = NewBuildLogger(trace, b.Log())
-	b.logger.Println(fmt.Sprintf("Running with %s\n  on %s (%s)", AppVersion.Line(), b.Runner.Name, b.Runner.ShortDescription()))
+	b.logger.Println("Running with", AppVersion.Line())
+	if b.Runner != nil && b.Runner.ShortDescription() != "" {
+		b.logger.Println("  on", b.Runner.Name, b.Runner.ShortDescription())
+	}
 
 	b.CurrentState = BuildRunStatePending